+++ src/AzzyBot.Bot/Commands/Autocompletes/AzuraCastMountAutocomplete.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
+using AzzyBot.Bot.Services;
 using AzzyBot.Bot.Services.Modules;
 using AzzyBot.Bot.Utilities.Records.AzuraCast;
 using AzzyBot.Core.Logging;
@@ -15,11 +16,12 @@ using Microsoft.Extensions.Logging;
 
 namespace AzzyBot.Bot.Commands.Autocompletes;
 
-public sealed class AzuraCastMountAutocomplete(ILogger<AzuraCastMountAutocomplete> logger, AzuraCastApiService azuraCast, DbActions dbActions) : IAutoCompleteProvider
+public sealed class AzuraCastMountAutocomplete(ILogger<AzuraCastMountAutocomplete> logger, AzuraCastApiService azuraCast, DbActions dbActions, DiscordBotService botService) : IAutoCompleteProvider
 {
     private readonly ILogger<AzuraCastMountAutocomplete> _logger = logger;
     private readonly AzuraCastApiService _azuraCast = azuraCast;
     private readonly DbActions _dbActions = dbActions;
+    private readonly DiscordBotService _botService = botService;
 
     public async ValueTask<IReadOnlyDictionary<string, object>> AutoCompleteAsync(AutoCompleteContext context)
     {
@@ -30,7 +32,7 @@ public sealed class AzuraCastMountAutocomplete(ILogger<AzuraCastMountAutocomplet
         if (stationId is 0)
             return new Dictionary<string, object>();
 
-        AzuraCastEntity? azuraCastEntity = await _dbActions.GetAzuraCastAsync(context.Guild.Id);
+        AzuraCastEntity? azuraCastEntity = await _dbActions.GetAzuraCastAsync(context.Guild.Id, loadPrefs: true);
         if (azuraCastEntity is null)
         {
             _logger.DatabaseAzuraCastNotFound(context.Guild.Id);
@@ -39,7 +41,21 @@ public sealed class AzuraCastMountAutocomplete(ILogger<AzuraCastMountAutocomplet
 
         string search = context.UserInput;
         string name = string.Empty;
-        AzuraStationRecord record = await _azuraCast.GetStationAsync(new(Crypto.Decrypt(azuraCastEntity.BaseUrl)), stationId);
+        AzuraStationRecord? record = null;
+        try
+        {
+            record = await _azuraCast.GetStationAsync(new(Crypto.Decrypt(azuraCastEntity.BaseUrl)), stationId);
+            if (record is null)
+            {
+                await _botService.SendMessageAsync(azuraCastEntity.Preferences.NotificationChannelId, $"I don't have the permission to access the **station** ({stationId}) endpoint.\n{AzuraCastApiService.AzuraCastPermissionsWiki}");
+                return new Dictionary<string, object>();
+            }
+        }
+        catch (InvalidOperationException)
+        {
+            return new Dictionary<string, object>();
+        }
+
         bool hlsAvailable = record.HlsUrl is not null;
         int maxMounts = (hlsAvailable) ? 24 : 25;
         Dictionary<string, object> results = new(25);