+++ src/AzzyBot.Bot/Commands/Autocompletes/AzuraCastSystemLogAutocomplete.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using AzzyBot.Bot.Services;
 using AzzyBot.Bot.Services.Modules;
 using AzzyBot.Bot.Utilities.Records.AzuraCast;
 using AzzyBot.Core.Logging;
@@ -13,11 +14,12 @@ using Microsoft.Extensions.Logging;
 
 namespace AzzyBot.Bot.Commands.Autocompletes;
 
-public sealed class AzuraCastSystemLogAutocomplete(ILogger<AzuraCastSystemLogAutocomplete> logger, AzuraCastApiService azuraCastApi, DbActions dbActions) : IAutoCompleteProvider
+public sealed class AzuraCastSystemLogAutocomplete(ILogger<AzuraCastSystemLogAutocomplete> logger, AzuraCastApiService azuraCastApi, DbActions dbActions, DiscordBotService botService) : IAutoCompleteProvider
 {
     private readonly ILogger<AzuraCastSystemLogAutocomplete> _logger = logger;
     private readonly AzuraCastApiService _azuraCastApi = azuraCastApi;
     private readonly DbActions _dbActions = dbActions;
+    private readonly DiscordBotService _botService = botService;
 
     public async ValueTask<IReadOnlyDictionary<string, object>> AutoCompleteAsync(AutoCompleteContext context)
     {
@@ -34,7 +36,13 @@ public sealed class AzuraCastSystemLogAutocomplete(ILogger<AzuraCastSystemLogAut
         string search = context.UserInput;
         Uri baseUrl = new(Crypto.Decrypt(azuraCast.BaseUrl));
         string apiKey = Crypto.Decrypt(azuraCast.AdminApiKey);
-        AzuraSystemLogsRecord systemLogs = await _azuraCastApi.GetSystemLogsAsync(baseUrl, apiKey);
+        AzuraSystemLogsRecord? systemLogs = await _azuraCastApi.GetSystemLogsAsync(baseUrl, apiKey);
+        if (systemLogs is null)
+        {
+            await _botService.SendMessageAsync(azuraCast.Preferences.NotificationChannelId, $"I don't have the permission to access the administrative **system logs** endpoint.\n{AzuraCastApiService.AzuraCastPermissionsWiki}");
+            return new Dictionary<string, object>();
+        }
+
         Dictionary<string, object> results = new(25);
         foreach (AzuraSystemLogEntryRecord log in systemLogs.Logs)
         {