+++ src/AzzyBot.Bot/Services/UpdaterService.cs
@@ -34,7 +34,7 @@ public sealed class UpdaterService(ILogger<UpdaterService> logger, AzzyBotSettin
             ["User-Agent"] = SoftwareStats.GetAppName
         };
 
-        string body = await _webService.GetWebAsync((isPreview) ? _previewUrl : _latestUrl, headers, true);
+        string? body = await _webService.GetWebAsync((isPreview) ? _previewUrl : _latestUrl, headers, true);
         if (string.IsNullOrWhiteSpace(body))
         {
             _logger.OnlineVersionEmpty();